Changed enum LanguageIdSeparator to LanguageIdStyle
LanguageIdSeparator was actually designed for BCP 47 language tag to
path/file name mapping. For now, we only support BCP 47 and a separator
variant, but technically, as long as we can support 1-to-1 mapping, it
does not need to be just separator variant. For this reason, I changed
the enum to LanguageIdStyle, and enum constants to BCP47 /
BCP47_UNDERSCORE.