Handle exception while parsing Java resource string
MessagePattern class throws an exception when a given string is not
parsed as a valid message format pattern string. Added exception
handling code in the code unescaping single quote in an input resource
string. When an exception was thrown by MessagePattern class, we don’t
need to unescape single quotes, because it’s not a message format
pattern.